refactoring + controller edits
