+++ src/main/java/be/thomasmore/qrace/controller/HomeController.java
@@ -4,10 +4,10 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class StartController {
+public class HomeController {
 
   @GetMapping("/")
   public String showStartPage() {
-    return "start";
+    return "index";
   }
 }
