+++ src/main/resources/templates/0-fragments/qrScanner.html
