+++ src/main/resources/templates/2-lobby/host-join.html
@@ -12,11 +12,11 @@
 
 <body>
     <!-- Help-knop en popup -->
-    <div th:replace="fragments\help-language :: help(helpText=${helpText})">
+    <div th:replace="0-fragments\help-language :: help(helpText=${helpText})">
     </div>
 
     <!-- Taal-knop -->
-    <div th:replace="fragments\help-language :: language"></div>
+    <div th:replace="0-fragments\help-language :: language"></div>
 
     <h1>Race Your Rivals</h1>
     <div class="button-container">
