+++ src/main/resources/templates/2-lobby/host.html
