+++ src/main/resources/templates/2-lobby/join.html
